requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
backrefs==6.1
beautifulsoup4==4.14.3
black==25.12.0
bleach==6.3.0
blinker==1.9.0
blis==1.3.3
cachetools==6.2.4
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
coverage==7.13.0
cycler==0.12.1
cymem==2.0.13
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.1
flake8==7.3.0
fonttools==4.61.1
fqdn==1.5.1
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.15
idna==3.11
ImageIO==2.37.2
imbalanced-learn==0.14.1
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
isort==7.0.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
langchain==1.2.0
langchain-core==1.2.4
langchain-openai==1.1.6
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.0
lark==1.3.1
lazy_loader==0.4
librt==0.7.4
lightgbm==4.6.0
lime==0.2.0.1
llvmlite==0.46.0
loguru==0.7.3
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mccabe==0.7.0
mergedeep==1.3.4
mistune==3.1.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.7.1
mkdocs-material-extensions==1.3.1
murmurhash==1.0.15
mypy==1.19.1
mypy_extensions==1.1.0
narwhals==2.14.0
nbclient==0.10.3
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
nodeenv==1.10.0
notebook==7.5.1
notebook_shim==0.2.4
numba==0.63.1
numpy==2.3.5
openai==2.14.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
paginate==0.5.7
pandas==2.3.3
pandera==0.27.1
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
pluggy==1.6.0
pre_commit==4.5.1
preshed==3.0.12
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyflakes==3.4.0
Pygments==2.19.2
pymcdm==1.3.1
pymdown-extensions==10.19.1
pyparsing==3.3.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-json-logger==4.0.0
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
pyyaml_env_tag==1.1
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
shap==0.50.0
six==1.17.0
sklearn-compat==0.1.5
slicer==0.0.8
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
stack-data==0.6.3
streamlit==1.52.2
streamlit-aggrid==1.2.1
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
thinc==8.3.10
threadpoolctl==3.6.0
tifffile==2025.12.20
tiktoken==0.12.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.4
typer-slim==0.20.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.2
uuid_utils==0.12.0
virtualenv==20.35.4
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.14
weasel==0.4.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
wrapt==2.0.1
xgboost==3.1.2
xxhash==3.6.0
zstandard==0.25.0
